refactor: Activity 도메인 컴포넌트 정리
