+++ src/domain/activity/components/display/ActivityDetailClient/ActivityHeader.tsx
