+++ src/domain/activity/components/display/ActivityDetailClient/ParticipantCounter.tsx
