+++ src/domain/activity/components/display/IntroSection/HeroBanner.tsx
