+++ src/domain/activity/components/display/SearchResult/SearchResultDataWrapper.tsx
