+++ src/domain/activity/components/display/SearchResultSkeleton.tsx
