+++ src/domain/activity/components/management/RegistrationForm/BasicInfoFields.tsx
