+++ src/domain/activity/components/management/RegistrationForm/ClientConfirmModal.tsx
