+++ src/domain/activity/components/management/RegistrationForm/DatePickerField.tsx
