+++ src/domain/activity/components/management/RegistrationForm/ImageUploader.tsx
