+++ src/domain/activity/components/management/RegistrationForm/hooks/useBannerImageUpload.ts
