+++ src/domain/activity/components/management/RegistrationForm/hooks/useIntroImageUpload.ts
