+++ src/domain/activity/components/management/RegistrationForm/hooks/useLeaveGuard.ts
