+++ src/domain/activity/components/management/RegistrationForm/utils/createEmptyTimeSlot.ts
