+++ src/domain/activity/components/management/RegistrationForm/utils/normalizeSubImages.ts
