+++ src/domain/activity/components/management/RegistrationForm/utils/validateRegistration.ts
