+++ src/domain/activity/components/management/types.ts
@@ -1,12 +1,3 @@
-export interface ActivityCardProps {
-  id: number;
-  title: string;
-  rating: number;
-  reviewCount: number;
-  price: number;
-  bannerImageUrl: string;
-  onDelete?: () => void;
-}
 export interface FormValues {
   title: string;
   category: string;
