--- src/domain/myactivities/components/type.ts
@@ -1,9 +1,0 @@
-export interface ActivityCardProps {
-  id: number;
-  title: string;
-  rating: number;
-  reviewCount: number;
-  price: number;
-  bannerImageUrl: string;
-  onDelete?: () => void;
-}
